services/intelli-search/requirements.txt
# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.9.2
pydantic-settings==2.5.2

# Search Engine
opensearch-py==2.3.1
opensearch-dsl==2.1.0

# AI/LLM Integrations
openai==1.57.0
anthropic==0.43.0             # Claude API
langchain==0.3.7
langchain-openai==0.2.9       # ChatOpenAI + OpenAI tool-calling agent
langchain-community==0.3.7    # StructuredTool helpers
langchain-core==0.3.29        # Prompts, runnables
tavily-python==0.5.0          # Tavily web search SDK
instructor==1.5.2             # Structured LLM outputs
sentence-transformers==2.7.0  # Embeddings
# torch is installed separately in Docker (CPU-only) — do not pin here
huggingface-hub==0.25.2

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
redis==5.0.1

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Logging & Observability
structlog==23.2.0
python-json-logger==2.0.7
pyyaml==6.0.1

opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
setuptools<82                                    # pkg_resources required by OTel instrumentation

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Async & Concurrency
aiohttp==3.9.1
asyncpg==0.29.0

# Utilities
requests==2.31.0
tenacity==8.5.0  # Retry logic
python-dateutil==2.8.2
